requirements.txt
acme==1.21.0
appdirs==1.4.4
apturl==0.5.2
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certbot==1.21.0
certbot-nginx==1.21.0
certifi==2020.6.20
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.4
ConfigArgParse==1.5.3
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
cssselect2==0.7.0
cupshelpers==1.0
cycler==0.11.0
daphne==4.0.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2.4
django-channels-jwt-auth-middleware==1.0.0
django-cors-headers==4.3.1
django-currentuser==0.7.0
django-jazzmin==2.6.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
duplicity==0.8.21
fasteners==0.14.1
filelock==3.8.0
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
GDAL==3.4.1
gyp==0.1
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.4
ifaddr==0.1.7
importlib-metadata==4.6.4
incremental==22.10.0
inflection==0.5.1
ipython_genutils==0.2.0
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.0.3
josepy==1.10.0
jsonschema==3.2.0
jupyter-core==4.9.1
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.5.1
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.5
nbformat==5.1.3
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
openapi-codec==1.3.2
openva-pipeline==0.9.2
oscrypto==1.3.0
OWSLib==0.25.0
packaging==21.3
pandas==2.0.3
paramiko==2.9.3
parsedatetime==2.6
pbr==5.8.0
pexpect==4.8.0
Pillow==10.0.0
platformdirs==2.5.2
plotly==5.4.0
ply==3.11
protobuf==3.12.4
psycopg2==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
PyChromecast==9.4.0
pycparser==2.21
pycrossva==0.94
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
PyICU==2.8.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==2.4.7
pypdf==4.1.0
pypng==0.20220715.0
pyproj==3.3.0
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyrsistent==0.18.1
pysqlcipher3==1.2.0
python-apt==2.4.0+ubuntu3
python-bidi==0.4.2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pythran==0.10.0
pytz==2023.3
pyxdg==0.27
PyYAML==6.0.1
qrcode==7.4.2
QScintilla==2.11.6
redis==5.0.0
reportlab==4.0.9
requests==2.31.0
requests-toolbelt==0.9.1
scipy==1.8.0
SecretStorage==3.3.1
service-identity==23.1.0
simplejson==3.19.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.4
svglib==1.5.1
sympy==1.9
systemd-python==234
tenacity==6.3.1
tinycss2==1.2.1
traitlets==5.1.1
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uritemplate==4.1.1
uritools==4.0.2
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.16.5
wadllib==1.3.6
webencodings==0.5.1
xdg==5
xhtml2pdf==0.2.15
xkit==0.0.0
zeroconf==0.38.3
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==6.0
